+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [2.3.0] - 2021-01-18
+## [2.3.0] - 2021-01-26
 ### Added
 - Integrate API apps into the CLI. Support API message transmission, reception,
   and the demo-rx app directly from `blocksat-cli` commands (`api send`, `api
@@ -20,6 +20,8 @@
   every successfully decrypted API message.
 - Option `--sender` on the API listener app to filter digitally signed messages
   from a selected sender.
+- Option `--channel` on the API listener and demo-rx apps to support
+  multiplexing and filtering of independent API packet streams.
 - Support for clearsigning and verification of plaintext messages.
 - Wait mechanism on the API sender app to wait until the API message is
   successfully transmitted over satellite, optionally disabled using